Overhaul ets2, search components to be powered by Pathom3
Need to do the same for ATS eventually.

This also came with a reworking of binary SII parsing to return a
Clojure map more directly. It would probably be a great fit for Pathom
itself, which could save a lot of wasted work parsing parts we don't
care about. But it's fast enough to be getting on with, and the Pathom
setup will cache any file that's already been parsed.