Use the SCS file parsing for parsing most core data
This improves on hard-coding it, at the cost of a bit of startup time.

There's the skeleton of parsing cargoes too, but it needs finishing and
testing. But I couldn't see a way to track which ones were "heavy"
according to the Heavy Cargo Pack for ATS.

This needs a bit of cleanup and to be refactored to shared code with
ETS2.